+++ curry_test.ts
@@ -1,7 +1,8 @@
-// Copyright 2021-present the Curry authors. All rights reserved. MIT license.
+// Copyright 2023-latest Tomoki Miyauchi. All rights reserved. MIT license.
+
 import { Pop, Shift, UnionFactorial } from "./curry.ts";
 import { curry } from "./curry.ts";
-import { assertEquals, assertEqualsTypes } from "./dev_deps.ts";
+import { assertEquals, assertEqualsTypes } from "./_dev_deps.ts";
 
 const arity0 = () => true;
 const arity1 = (a: unknown) => a;