+++ dev_deps.ts
@@ -1,3 +1,0 @@
-// Copyright 2021-present the Curry authors. All rights reserved. MIT license.
-export { assertEquals } from "https://deno.land/std@0.97.0/testing/asserts.ts";
-export const assertEqualsTypes = <T, U extends T = T>(_actual?: U): void => {};