+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# [1.1.0-beta.3](https://github.com/TomokiMiyauci/curry/compare/1.1.0-beta.2...1.1.0-beta.3) (2023-06-05)
+
+
+### Features
+
+* **partial:** rename partial functions ([ac74154](https://github.com/TomokiMiyauci/curry/commit/ac7415418a28b8c70a7c80b1bb44c00f613e1fa2))
+
+# [1.1.0-beta.2](https://github.com/TomokiMiyauci/curry/compare/1.1.0-beta.1...1.1.0-beta.2) (2023-06-05)
+
+
+### Bug Fixes
+
+* **partial:** rename return function name ([2ab4f3b](https://github.com/TomokiMiyauci/curry/commit/2ab4f3b806a117b8ec9ae4ea709d81196b256575))
+
+# [1.1.0-beta.1](https://github.com/TomokiMiyauci/curry/compare/v1.0.0...1.1.0-beta.1) (2023-06-05)
+
+
+### Features
+
+* **partial:** add partial function ([3379335](https://github.com/TomokiMiyauci/curry/commit/3379335ca41eeacddb29218edf92b8c5e34c39fa))
+* **partial:** add partial right function ([46ed432](https://github.com/TomokiMiyauci/curry/commit/46ed43287a34b6ef7165c7db95cec08b74ac1f8a))
+* **partial:** add partial tail function ([96b1d7b](https://github.com/TomokiMiyauci/curry/commit/96b1d7ba8390328100a0f46465fb53e5d847c3af))
+
 # v1.0.0 (2021-06-15)
 
 ## ✨ New Features